refactor: animate 를 timeline 으로 변경
