+++ src/apps/docs/features/play/components/PlayTest1/PlayTest1.tsx
@@ -14,7 +14,7 @@ function PlayTest1({ className }: PlayTest1Props) {
   useEffect(() => {
     if (box1.current) {
       paliga
-        .animate([box1.current], {
+        .timeline([box1.current], {
           x: 200,
           duration: 1000,
         })
@@ -22,7 +22,7 @@ function PlayTest1({ className }: PlayTest1Props) {
     }
 
     if (box2.current) {
-      paliga2.animate([box2.current], {
+      paliga2.timeline([box2.current], {
         x: 200,
         duration: 1000,
       });
