+++ src/core/Paliga.ts
@@ -1,11 +1,9 @@
 import { CSSProperties } from "react";
 import { animationsRunner, getAnimationToStyle } from "../helpers/animationHelpers";
-import { scheduleRunner } from "../helpers/scheduleHelpers";
+import { getSchedule, scheduleRunner } from "../helpers/scheduleHelpers";
 import { getInnerHeight } from "../helpers/styleHelpers";
 import {
   TAnimateOptions,
-  TAnimation,
-  TAnimationState,
   TIntersectionPlayOptions,
   TPlayOptions,
   TSchedule,
@@ -19,6 +17,7 @@ import { throttle } from "./timeUtils";
 
 export class Paliga {
   #state: "idle" | "running" | "paused" = "idle";
+  #progress: number = 0;
   #totalDuration: number = 0;
   #animates: ({ elements: HTMLElement[] } & TAnimateOptions)[] = [];
   #segments: TSegment[] = [];
@@ -31,7 +30,7 @@ export class Paliga {
   }
 
   /** 연속 실행기 */
-  play({ iteration, onAllAnimationEnd }: TPlayOptions = {}) {
+  play({ iteration, onAnimationEnd, onAllAnimationEnd }: TPlayOptions = {}) {
     let newIteration = iteration;
 
     if (this.#state === "running") {
@@ -40,36 +39,79 @@ export class Paliga {
     this.#state = "running";
 
     /** 모든 애니메이션 종료 시 콜백 */
-    const handleAllAnimationEnd: TPlayOptions["onAllAnimationEnd"] = ({ segments }) => {
+    const allAnimationEnd = () => {
       // # 반복
       if (newIteration !== undefined && newIteration > 1 && this.#state !== "paused") {
         if (newIteration !== Infinity) {
           newIteration--;
         }
 
-        this.#runner({
-          schedules: this.#schedule,
-          onAllAnimationEnd: handleAllAnimationEnd,
+        this.#progress = 0;
+
+        scheduleRunner({
+          startProgress: this.#progress,
+          endProgress: 1,
+          scheduleList: this.#schedule,
+          onAnimationEnd: (data) => {
+            const { index } = data;
+
+            if (onAnimationEnd) {
+              onAnimationEnd(data);
+            }
+
+            if (index < this.#schedule.length - 1) {
+              return;
+            }
+
+            allAnimationEnd();
+          },
         });
         return;
       }
 
       this.#state = "idle";
+      this.#progress = 0;
 
       // # 모든 애니메이션 종료 시 호출
       if (
         typeof onAllAnimationEnd === "function" &&
         (newIteration === undefined || newIteration <= 1)
       ) {
-        onAllAnimationEnd({ segments: segments ? [...segments] : segments });
+        onAllAnimationEnd({ segments: this.#segments });
       }
     };
 
-    // 실행
-    this.#runner({
-      schedules: this.#schedule,
-      onAllAnimationEnd: handleAllAnimationEnd,
+    scheduleRunner({
+      startProgress: this.#progress,
+      endProgress: 1,
+      scheduleList: this.#schedule,
+      onFrame: ({ progress }) => {
+        this.#progress = progress;
+
+        if (this.#state === "paused") {
+          return false;
+        }
+      },
+      onAnimationEnd: (data) => {
+        const { index } = data;
+
+        if (onAnimationEnd) {
+          onAnimationEnd(data);
+        }
+
+        if (index < this.#schedule.length - 1) {
+          return;
+        }
+
+        allAnimationEnd();
+      },
     });
+
+    // 실행
+    // this.#runner({
+    //   schedules: this.#schedule,
+    //   onAllAnimationEnd: allAnimationEnd,
+    // });
   }
 
   /** 교차 시 실행 */
@@ -121,11 +163,13 @@ export class Paliga {
 
   /** progress 에 따른 스타일 적용 */
   progress(progress: number) {
-    this.#eachSchedule((schedule) => {
-      this.#scheduleProgress({
-        startProgress: progress,
-        schedule,
-      });
+    this.#state = "paused";
+    this.#progress = progress;
+
+    scheduleRunner({
+      startProgress: progress,
+      endProgress: progress,
+      scheduleList: this.#schedule,
     });
   }
 
@@ -228,7 +272,7 @@ export class Paliga {
         const scrollY = getScrollY(newRoot);
         const diff = newEnd - newStartY;
         const progress = limitRange((newTrigger + scrollY - newStartY) / diff);
-        console.log("> p: ", prevProgress.toFixed(2), progress.toFixed(2));
+        // console.log("> p: ", prevProgress.toFixed(2), progress.toFixed(2));
 
         scrollRunner({
           startProgress: prevProgress,
@@ -247,17 +291,13 @@ export class Paliga {
   /** 애니메이션 중단 */
   pause() {
     this.#state = "paused";
-
-    this.#eachSchedule((schedule) => {
-      schedule.state = "paused";
-    });
   }
 
   /** 애니메이션 정의 */
-  animate(elements: HTMLElement[], options: TAnimateOptions): Paliga;
-  animate(options: TAnimateOptions): Paliga;
+  timeline(elements: HTMLElement[], options: TAnimateOptions): Paliga;
+  timeline(options: TAnimateOptions): Paliga;
   // eslint-diabled-next-line no-dupe-class-members
-  animate(elements: HTMLElement[] | TAnimateOptions, options?: TAnimateOptions) {
+  timeline(elements: HTMLElement[] | TAnimateOptions, options?: TAnimateOptions) {
     const newOptions = options ?? elements;
     let newElements: HTMLElement[];
 
@@ -286,7 +326,7 @@ export class Paliga {
       elements: newElements,
     });
 
-    const { style, ...otherOptions } = newOptions;
+    const { beforeStyle, ...otherOptions } = newOptions;
 
     // # 엘리먼트별 트랜지션 계산
     newElements.forEach((element, i) => {
@@ -295,7 +335,7 @@ export class Paliga {
         : this.#segments.length;
       let newTransition: TTransition = {
         ...otherOptions,
-        animateIndex: this.#animates.length - 1,
+        // animateIndex: this.#animates.length - 1,
         element,
         easing: otherOptions.easing ?? "linear",
         direction: otherOptions.direction ?? "normal",
@@ -308,7 +348,8 @@ export class Paliga {
       element.dataset["traniId"] = String(id);
 
       // # 엘리먼트에 스타일 적용
-      if (style) {
+      if (beforeStyle) {
+        const style = beforeStyle(i);
         let k: keyof CSSProperties;
         for (k in style) {
           const newK = k as unknown as number;
@@ -373,7 +414,7 @@ export class Paliga {
       this.#totalDuration = Math.max(this.#segments[id].maxDuration, this.#totalDuration);
     }, []);
 
-    this.#schedule = this.#toSchedule({
+    this.#schedule = getSchedule({
       totalDuration: this.#totalDuration,
       segments: this.#segments,
     });
@@ -407,122 +448,6 @@ export class Paliga {
     return this;
   }
 
-  /** 스케쥴을 생성 */
-  #toSchedule({
-    totalDuration,
-    segments,
-  }: {
-    totalDuration: number;
-    segments: TSegment[];
-  }): TSchedule[] {
-    const len = segments.length;
-    const schedule: TSchedule[] = [];
-    let k = 0;
-
-    // # 세그먼트 순회
-    while (k < len) {
-      const segment = segments[k];
-      const curSchedule: TSchedule = {
-        id: segment.id,
-        state: "idle",
-        maxDuration: segment.maxDuration,
-        progress: 0,
-        element: segment.element,
-        animations: [],
-      };
-      const animations = curSchedule.animations;
-      const { transitions } = segment;
-      const tLen = transitions.length;
-      let prevAnimation: TAnimation | undefined = undefined;
-      let j = 0;
-
-      // # 트랜지션 순회
-      while (j < tLen) {
-        const transition = transitions[j];
-        const curDelay = transition.delay ?? 0;
-        const sumDuration = transition.duration + curDelay;
-
-        // # 시작점
-        const fromDuration = prevAnimation ? prevAnimation.to.duration : 0;
-        const from: TAnimationState = {
-          duration: fromDuration,
-          progress: prevAnimation ? prevAnimation.to.progress : 0,
-          x: prevAnimation?.to.x ?? (transition.x !== undefined ? 0 : undefined),
-          y: prevAnimation?.to.y ?? (transition.y !== undefined ? 0 : undefined),
-          opacity:
-            prevAnimation?.to.opacity ??
-            (transition.opacity !== undefined
-              ? Number(transition.element.style.opacity ?? 1)
-              : undefined),
-        };
-
-        // # 딜레이
-        const toDelay: TAnimation["toDelay"] = transition.delay
-          ? {
-              duration: fromDuration + curDelay,
-              progress: (fromDuration + curDelay) / totalDuration,
-            }
-          : undefined;
-
-        // # 종료
-        const toDuration = toDelay
-          ? toDelay.duration + transition.duration
-          : from.duration + transition.duration;
-        const to: TAnimationState = {
-          duration: toDuration,
-          progress: toDuration / totalDuration,
-          x: transition.x
-            ? transition.direction === "reverse"
-              ? -transition.x + (from.x ?? 0)
-              : transition.x + (from.x ?? 0)
-            : transition.x,
-          y: transition.y
-            ? transition.direction === "reverse"
-              ? -transition.y + (from.y ?? 0)
-              : transition.y + (from.y ?? 0)
-            : transition.y,
-          opacity: transition.opacity,
-        };
-
-        const newTo: TAnimationState = {
-          duration: to.duration,
-          progress: to.progress,
-          x: to.x === undefined && from.x !== undefined ? from.x : to.x,
-          y: to.y === undefined && from.y !== undefined ? from.y : to.y,
-          opacity:
-            to.opacity === undefined && from.opacity !== undefined ? from.opacity : to.opacity,
-        };
-
-        // 스케쥴
-        const newAnimation: TAnimation = {
-          sumDuration,
-          from,
-          toDelay,
-          to: newTo,
-          element: transition.element,
-          direction: transition.direction,
-          duration: transition.duration,
-          // fillMode: transition.fillMode,
-          delay: transition.delay,
-          easing: transition.easing,
-          onFrame: transition.onFrame,
-        };
-
-        prevAnimation = newAnimation;
-
-        animations.push(newAnimation);
-
-        j++;
-      }
-
-      schedule.push(curSchedule);
-
-      k++;
-    }
-
-    return schedule;
-  }
-
   /** 지정된 스케쥴을 progress에 따른 애니메이션 적용 */
   #scheduleProgress({
     startProgress = 0,
@@ -537,7 +462,7 @@ export class Paliga {
     schedule: TSchedule;
     onAnimationEnd?: (params: { progress: number }) => void;
   }) {
-    const { maxDuration, animations, element } = schedule;
+    const { maxDuration, animations } = schedule;
     const elapsed = maxDuration * startProgress;
     const len = animations.length;
     let i = 0;
@@ -565,114 +490,6 @@ export class Paliga {
       animations,
       onAnimationEnd,
     });
-
-    // const filteredAnimations = rangeAnimations({ startProgress, endProgress, animations });
-    // const animation = filteredAnimations.shift();
-    // console.log("> ", maxDuration);
-
-    // const animation = animations[animationIndex];
-    // const isNormal = startProgress < endProgress;
-    // const newStartProgress = isNormal ? startProgress : endProgress;
-    // const newEndProgress = isNormal ? endProgress : startProgress;
-    // const diff = newEndProgress - newStartProgress;
-
-    // framer({
-    //   duration: duration ?? animation.duration,
-    //   startProgress: newStartProgress,
-    //   endProgress: newEndProgress,
-    //   onFrame: ({ progress }) => {
-    //     const newProgress =
-    //       isNormal || diff === 0
-    //         ? progress
-    //         : Math.max(
-    //             diff * (1 - (progress - newStartProgress) / diff) + newStartProgress,
-    //             newStartProgress,
-    //           );
-
-    //     const styles = getAnimationToStyle({
-    //       maxDuration,
-    //       progress: newProgress,
-    //       animation,
-    //     });
-    //     const { x, y, opacity } = styles;
-
-    //     ApplyStyles.router({ el: element, x, y, opacity });
-    //   },
-    //   onDone: (data) => {
-    //     if (typeof onAnimationEnd === "function") {
-    //       onAnimationEnd(data);
-    //     }
-    //   },
-    // });
-  }
-
-  /** 전체 progress 를 기준으로 지정된 animation 스타일 정보를 반환  */
-  // #progressAnimationStyle({
-  //   maxDuration,
-  //   progress,
-  //   animation,
-  // }: {
-  //   maxDuration: number;
-  //   progress: number;
-  //   animation: TAnimation;
-  // }): TStylesState & { animationProgress: number } {
-  //   const { from, toDelay, to, easing = "linear", onFrame } = animation;
-  //   const startProgress = (toDelay ? toDelay.progress : from.progress) ?? 0;
-  //   const animationProgress = Math.min(this.#calcProgress(progress, startProgress, to.progress), 1);
-  //   const isDelay = animationProgress < 0;
-  //   const frame = onFrame
-  //     ? onFrame({ maxDuration, progress, animationProgress, animation })
-  //     : undefined;
-  //   const easingProgress =
-  //     frame?.progress ??
-  //     Easing.router({
-  //       progress: animationProgress,
-  //       easing,
-  //     });
-
-  //   let x: number | undefined = from.x;
-  //   let y: number | undefined = from.y;
-  //   let opacity: number | undefined = from.opacity;
-
-  //   if (!isDelay) {
-  //     x = frame?.x ?? this.#calcValue(from.x, to.x, easingProgress);
-  //     y = frame?.y ?? this.#calcValue(from.y, to.y, easingProgress);
-  //     opacity = frame?.opacity ?? this.#calcValue(from.opacity, to.opacity, easingProgress);
-  //   }
-
-  //   return {
-  //     animationProgress,
-  //     x,
-  //     y,
-  //     opacity,
-  //   };
-  // }
-
-  /** 지정된 구간의 진행도를 반환 */
-  // #calcProgress = (progress: number, startProgress: number, endProgress: number) => {
-  //   return (progress - startProgress) / (endProgress - startProgress);
-  // };
-
-  /** progres에 따른 값을 계산 */
-  #calcValue(fromValue?: number, value?: number, progress: number = 0, isReverse = false) {
-    if (value === undefined && fromValue === undefined) {
-      return;
-    }
-    if (fromValue !== undefined && value === undefined) {
-      return fromValue;
-    }
-    if (fromValue === value) {
-      return value;
-    }
-
-    const newFromValue = fromValue ?? 0;
-    const newValue = value ?? 0;
-
-    if (isReverse) {
-      return (newValue - newFromValue) * -progress + newFromValue;
-    }
-
-    return (newValue - newFromValue) * progress + newFromValue;
   }
 
   /** 스케쥴 엘리먼트 순회 */
@@ -741,46 +558,5 @@ export class Paliga {
         }
       },
     });
-  }
-
-  /** 전체 스케쥴 실행기 */
-  #runner({
-    schedules,
-    onAllAnimationEnd,
-  }: { schedules: TSchedule[] } & Pick<TPlayOptions, "onAllAnimationEnd">) {
-    const len = schedules.length;
-    let i = 0;
-    let doneCount = 0;
-
-    // console.log("> runner: ", this.#schedule);
-
-    while (i < len) {
-      const schedule = schedules[i] ?? {};
-
-      if (schedule.state === "running") {
-        return;
-      }
-
-      // # 일시 중지가 아닐경우는 초기화
-      if (schedule.state !== "paused") {
-        schedule.progress = 0;
-      }
-
-      schedule.state = "running";
-
-      this.#scheduleRunner({
-        startProgress: schedule.progress,
-        schedule,
-        onAnimationEnd: () => {
-          ++doneCount;
-
-          if (typeof onAllAnimationEnd === "function" && doneCount === len) {
-            onAllAnimationEnd({ segments: this.#segments });
-          }
-        },
-      });
-
-      i++;
-    }
   }
 }
